Load extensions after other service providers
And let them know what kind of request this is via a “type” binding on
the container. Not sure if there is a better way to do this. But they
need to know somehow, so extenders can act selectively (e.g. the
ForumClient extender should only act on forum requests, not on API or
admin requests)