Remove redundant installation info lines
Why do we even have this ~~lever~~ line? 

Looks like a bit of oversight when we consolidated docs and updated for stable. Probably made sense at the time, now we have installation docs that we keep up to date. Remove this line to prevent confusion because it's out of date.